+++ src/test/java/racingcar/domain/TryCntTest.java
@@ -1,29 +1,20 @@
 package racingcar.domain;
 
-import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
+import static org.assertj.core.api.Assertions.assertThat;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import racingcar.exception.trycnt.TryCntIsNumberException;
-import racingcar.exception.trycnt.TryCntIsPositiveException;
 
 class TryCntTest {
-
     @Test
-    @DisplayName("[에러] 시도 횟수가 숫자가 아니라면 IllegalArgumentException 발생")
-    public void checkTryCntIsNumber() {
-        String tryCnt = "일";
-        assertThatExceptionOfType(TryCntIsNumberException.class)
-                .isThrownBy(() -> TryCnt.checkTryCount(tryCnt))
-                .withMessageMatching("시도 횟수는 숫자여야 합니다.");
-    }
-
-    @Test
-    @DisplayName("[에러] 시도 횟수가 음수라면 IllegalArgumentException 발생")
-    public void checkTryCntIsPositive() {
-        String tryCnt = "-1";
-        assertThatExceptionOfType(TryCntIsPositiveException.class)
-                .isThrownBy(() -> TryCnt.checkTryCount(tryCnt))
-                .withMessageMatching("시도 횟수는 양수여야 합니다.");
+    @DisplayName("시도 횟수 객체 생성")
+    public void createTryCntInstance() {
+        // given
+        String input = "1";
+        // when
+        TryCnt tryCnt = TryCnt.createTryCnt(input);
+        // then
+        assertThat(tryCnt).isNotNull();
+        assertThat(tryCnt).isInstanceOf(TryCnt.class);
     }
 }
